Decompose DashboardPage into hook plus toolbar plus widget registry
Was a single 509-line file. Now:
  - hooks/useDashboardStats.js owns data loading and the entire stats
    aggregation pipeline (one effect, one pure buildStats function with
    small named helpers); exposes { stats, allChallenges, error }
  - components/DashboardToolbar.jsx owns the title row and the
    customize/save/cancel/widgets controls
  - components/widgetRenderer.jsx owns the widget-id -> component
    registry so adding a widget is a one-line change
  - DashboardPage.jsx (now 178 lines) just orchestrates state and
    renders banners, the grid, the catalog, and snackbar

Also: replaces the setState-in-effect that auto-cleared comparison mode
with state updates in the toggle/clear handlers (silences the new
react-hooks rule and matches React 19 guidance). Behavior unchanged.

Refs #42.
